List View: Unify shortcut handlers
